This project is a repository created for the Restaurant ordering and information website using HTML, CSS and JavaScript.
This project includes a website developed using HTML, CSS and JavaScript. It offers a user-friendly interface that allows you to order food from a restaurant. There is also informative content about the restaurant. This repository provides an excellent starting point for developing web development skills and offers a great opportunity to practice.